Cleaning the back end and making it dev ready shortly
