--- back/models/recruiter.js
@@ -1,72 +1,0 @@
-module.exports = function(sequelize, Sequelize) {
-
-    var Recruiter = sequelize.define('recruiter', {
-        // id: {
-        //     autoIncrement: true,
-        //     primaryKey: true,
-        //     type: Sequelize.INTEGER
-        // },
-        //
-        // firstname: {
-        //     type: Sequelize.STRING,
-        //     allowNull: false,
-        //     validate: {
-        //         notEmpty: true
-        //     }
-        // },
-        //
-        // lastname: {
-        //     type: Sequelize.STRING,
-        //     allowNull: false,
-        //     validate: {
-        //         notEmpty: true
-        //     }
-        // },
-
-        username: {
-            type: Sequelize.STRING,
-            allowNull: true
-            // validate: {
-            //     notEmpty: true
-            // }
-        },
-
-        password: {
-            type: Sequelize.STRING,
-            allowNull: false
-            // validate: {
-            //     notEmpty: true
-            // }
-        }
-
-        // email: {
-        //     type: Sequelize.STRING,
-        //     validate: {
-        //         isEmail: true
-        //     }
-        // },
-        //
-        // company: {
-        //     type: Sequelize.STRING,
-        //     allowNull: true,
-        //     validate: {
-        //         notEmpty: true
-        //     }
-        // },
-        //
-        // phone: {
-        //     type: Sequelize.STRING,
-        //     allowNull: true,
-        //     validate: {
-        //         notEmpty: true,
-        //         len: [10]
-        //     }
-        // },
-        //
-        // photo: {
-        //     type: Sequelize.STRING
-        // }
-    });
-
-    return Recruiter;
-};
